refactor(pm): split package cache layer out of downloader
Untangle the three concerns that were mixed in downloader.rs:
- downloader.rs: pure network — download_bytes + client + is_git_url/is_registry_tarball_url.
- package_cache.rs (new): cache-path layout, seeded-slot lookups, registry_cache_*,
  extract_to_cache, download_and_extract_to_cache, ExtractOutcome — owns the _resolved contract.
- extractor.rs: unchanged raw gzip/tar primitive (extract_and_write), now called only by package_cache.

Pure move + re-import, no behavior change.
